+++ src/pages/home/homePage/Hero.js
